Misc: Fix unused parameter warnings. Fix sign conversion warnings. Update lib/Str with warning fxes.

+++ libs/Str/Str.h
@@ -226,7 +226,7 @@ void    Str::set(const char* src)
     int buf_len = (int)strlen(src)+1;
     if (Capacity < buf_len)
         reserve_discard(buf_len);
-    memcpy(Data, src, buf_len);
+    memcpy(Data, src, (size_t)buf_len);
     Owned = 1;
 }
 
@@ -236,7 +236,7 @@ void    Str::set(const char* src, const char* src_end)
     int buf_len = (int)(src_end-src)+1;
     if ((int)Capacity < buf_len)
         reserve_discard(buf_len);
-    memcpy(Data, src, buf_len-1);
+    memcpy(Data, src, (size_t)(buf_len - 1));
     Data[buf_len-1] = 0;
     Owned = 1;
 }
@@ -246,7 +246,7 @@ void    Str::set(const Str& src)
     int buf_len = (int)strlen(src.c_str())+1;
     if ((int)Capacity < buf_len)
         reserve_discard(buf_len);
-    memcpy(Data, src.c_str(), buf_len);
+    memcpy(Data, src.c_str(), (size_t)buf_len);
     Owned = 1;
 }
 
@@ -256,7 +256,7 @@ void    Str::set(const std::string& src)
     int buf_len = (int)src.length()+1;
     if ((int)Capacity < buf_len)
         reserve_discard(buf_len);
-    memcpy(Data, src.c_str(), buf_len);
+    memcpy(Data, src.c_str(), (size_t)buf_len);
     Owned = 1;
 }
 #endif
@@ -439,11 +439,11 @@ void    Str::reserve(int new_capacity)
     else
     {
         // Disowned or LocalBuf -> Heap
-        new_data = (char*)STR_MEMALLOC(new_capacity * sizeof(char));
+        new_data = (char*)STR_MEMALLOC((size_t)new_capacity * sizeof(char));
     }
 
     // string in Data might be longer than new_capacity if it wasn't owned, don't copy too much
-    strncpy(new_data, Data, new_capacity - 1);
+    strncpy(new_data, Data, (size_t)new_capacity - 1);
     new_data[new_capacity - 1] = 0;
 
     if (Owned && !is_using_local_buf())
@@ -472,7 +472,7 @@ void    Str::reserve_discard(int new_capacity)
     else
     {
         // Disowned or LocalBuf -> Heap
-        Data = (char*)STR_MEMALLOC(new_capacity * sizeof(char));
+        Data = (char*)STR_MEMALLOC((size_t)new_capacity * sizeof(char));
         Capacity = new_capacity;
     }
     Owned = 1;
@@ -486,8 +486,8 @@ void    Str::shrink_to_fit()
     if (Capacity <= new_capacity)
         return;
 
-    char* new_data = (char*)STR_MEMALLOC(new_capacity * sizeof(char));
-    memcpy(new_data, Data, new_capacity);
+    char* new_data = (char*)STR_MEMALLOC((size_t)new_capacity * sizeof(char));
+    memcpy(new_data, Data, (size_t)new_capacity);
     STR_MEMFREE(Data);
     Data = new_data;
     Capacity = new_capacity;
@@ -511,13 +511,13 @@ int     Str::setfv(const char* fmt, va_list args)
     len = vsnprintf(Data, len + 1, fmt, args2);
 #else
     // First try
-    int len = vsnprintf(Owned ? Data : NULL, Owned ? Capacity : 0, fmt, args);
+    int len = vsnprintf(Owned ? Data : NULL, Owned ? (size_t)Capacity : 0, fmt, args);
     STR_ASSERT(len >= 0);
 
     if (Capacity < len + 1)
     {
         reserve_discard(len + 1);
-        len = vsnprintf(Data, len + 1, fmt, args2);
+        len = vsnprintf(Data, (size_t)len + 1, fmt, args2);
     }
 #endif
 
@@ -541,7 +541,7 @@ int     Str::setfv_nogrow(const char* fmt, va_list args)
     if (Capacity == 0)
         return 0;
 
-    int w = vsnprintf(Data, Capacity, fmt, args);
+    int w = vsnprintf(Data, (size_t)Capacity, fmt, args);
     Data[Capacity - 1] = 0;
     Owned = 1;
     return (w == -1) ? Capacity - 1 : w;
@@ -574,7 +574,7 @@ int     Str::append_from(int idx, const char* s, const char* s_end)
     int add_len = (int)(s_end - s);
     if (Capacity < idx + add_len + 1)
         reserve(idx + add_len + 1);
-    memcpy(Data + idx, (const void*)s, add_len);
+    memcpy(Data + idx, (const void*)s, (size_t)add_len);
     Data[idx + add_len] = 0; // Our source data isn't necessarily zero-terminated
     STR_ASSERT(Owned);
     return add_len;
@@ -598,13 +598,13 @@ int     Str::appendfv_from(int idx, const char* fmt, va_list args)
     add_len = vsnprintf(Data + idx, add_len + 1, fmt, args2);
 #else
     // First try
-    int add_len = vsnprintf(Owned ? Data + idx : NULL, Owned ? Capacity - idx : 0, fmt, args);
+    int add_len = vsnprintf(Owned ? Data + idx : NULL, Owned ? (size_t)(Capacity - idx) : 0, fmt, args);
     STR_ASSERT(add_len >= 0);
 
     if (Capacity < idx + add_len + 1)
     {
         reserve(idx + add_len + 1);
-        add_len = vsnprintf(Data + idx, add_len + 1, fmt, args2);
+        add_len = vsnprintf(Data + idx, (size_t)add_len + 1, fmt, args2);
     }
 #endif
