+++ test_engine/imgui_te_util.cpp
@@ -178,7 +178,7 @@ bool ImGui::InputText(const char* label, Str* str, ImGuiInputTextFlags flags, Im
     cb_user_data.StrObj = str;
     cb_user_data.ChainCallback = callback;
     cb_user_data.ChainCallbackUserData = user_data;
-    return InputText(label, (char*)str->c_str(), str->capacity() + 1, flags, InputTextCallbackStr, &cb_user_data);
+    return InputText(label, (char*)str->c_str(), (size_t)str->capacity() + 1, flags, InputTextCallbackStr, &cb_user_data);
 }
 
 bool ImGui::InputTextMultiline(const char* label, Str* str, const ImVec2& size, ImGuiInputTextFlags flags, ImGuiInputTextCallback callback, void* user_data)
@@ -190,5 +190,5 @@ bool ImGui::InputTextMultiline(const char* label, Str* str, const ImVec2& size, 
     cb_user_data.StrObj = str;
     cb_user_data.ChainCallback = callback;
     cb_user_data.ChainCallbackUserData = user_data;
-    return InputTextMultiline(label, (char*)str->c_str(), str->capacity() + 1, size, flags, InputTextCallbackStr, &cb_user_data);
+    return InputTextMultiline(label, (char*)str->c_str(), (size_t)str->capacity() + 1, size, flags, InputTextCallbackStr, &cb_user_data);
 }
