+++ test_engine/imgui_te_util.h
@@ -33,7 +33,7 @@ struct ImMovingAverage
     int             FillAmount;
 
     ImMovingAverage()               { Accum = (TYPE)0; Idx = FillAmount = 0; }
-    void    Init(int count)         { Samples.resize(count); memset(Samples.Data, 0, Samples.Size * sizeof(TYPE)); Accum = (TYPE)0; Idx = FillAmount = 0; }
+    void    Init(int count)         { Samples.resize(count); memset(Samples.Data, 0, (size_t)Samples.Size * sizeof(TYPE)); Accum = (TYPE)0; Idx = FillAmount = 0; }
     void    AddSample(TYPE v)       { Accum += v - Samples[Idx]; Samples[Idx] = v; if (++Idx == Samples.Size) Idx = 0; if (FillAmount < Samples.Size) FillAmount++;  }
     TYPE    GetAverage() const      { return Accum / (TYPE)FillAmount; }
     int     GetSampleCount() const  { return Samples.Size; }
